requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.1
opencv-python==4.2.0.32
pytube3==9.6.4
scenedetect==0.5.1.1
tqdm==4.43.0
typing-extensions==3.7.4.1
Werkzeug==1.0.0
